+++ anthem/lyrics/__init__.py
@@ -1,0 +1,16 @@
+# -*- coding: utf-8 -*-
+
+"""
+Lyrics are isolated pieces of code that you can use and reuse in
+your own songs.
+
+They can be structured as you desire, from Haikus to Odes as long as the songs
+you write reach the zen of Python and execute harmoniously.
+
+Examples of lyrics are:
+
+ * load a csv, xml or yaml file
+ * add a xmlid on a record
+ * upsert a record
+
+"""