+++ HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ---------------
 
+Unreleased
+++++++++++
+
+- Add a lyric to create a xmlid
+- Add a lyric to create or update a record
+
+
 0.1.1 (2016-06-23)
 ++++++++++++++++++